refactor: extract 3 aggregation-helper panels from PanelViews.tsx
Batch 5 of the per-panel split for issue #144: districts, blocks,
analytics - all lean on getDistrictStats/getBlockStats from usePanelData.

- New panels/DistrictsPanel.tsx ({ currentUser, schools, students,
  getDistrictStats } props) - moves in the expandedDist useState
  (previously top-level in PanelViews.tsx, confirmed used only here)
  and the district drill-down logic.
- New panels/BlocksPanel.tsx ({ currentUser, getBlockStats } props).
- New panels/AnalyticsPanel.tsx ({ currentUser, schools, students,
  getDistrictStats, getBlockStats } props - the widest prop set in
  this batch, since it branches by role to pick which aggregation to
  render).
- PanelViews.tsx's 3 if-blocks replaced with one-line router calls.
- Removed the now-orphaned top-level expandedDist useState (kept
  expandedDistRpt, a separate variable still used by the reports
  panel).
- Unused-import audit: no new dead imports this round - MapPin/
  SchoolIcon/Award/BarChart3 all confirmed still used elsewhere in the
  remaining file (reports/teachers panels).

Caught and fixed a genuine off-by-one during extraction: DistrictsPanel's
JSX initially dropped one closing </div> at the very end of the block
(cut at the second-to-last line instead of the last). Caught via a
structural re-read of the built file before running tsc, not by tsc
itself - worth flagging since it's a reminder that byte-range slicing
needs a boundary sanity check even after tsc later confirms it (tsc
would have failed loudly, but catching it earlier is better practice).

Verified: tsc --noEmit clean, vite build succeeds, byte-diff of every
moved block. Live click-through: Districts + Analytics as Punjab State
Coordinator (ADMIN role) - district drill-down (expandedDist) works,
Analytics correctly shows "Geographical Analytics" (the isAdmin
branch, using getDistrictStats). Re-checked Analytics as Superadmin -
correctly shows "Performance Analytics" instead (the non-admin
fallback branch, data = schools), confirming the role-conditional
title/data logic survived the move intact.
